requirements.txt
flask
flask-boto3
flask-login
flask_migrate
flask-restplus
flask_sqlalchemy
flask_wtf
jira
mysql-connector
python-dotenv
requests
sample-sheet
sevenbridges-python
smart_open
